Put tyrian converter in a separate package
