+++ src/main/scala/simer/html/converter/utils/package.scala
@@ -1,0 +1,12 @@
+package simer.html.converter
+
+package object utils {
+  def tripleQuote(string: String): String =
+    string.trim match {
+      case string if string.contains("\"") || string.contains("\n") || string.contains("\\") =>
+        s"""\"\"\"$string\"\"\""""
+
+      case string =>
+        s""""$string""""
+    }
+}
